Revert "Existing bl and db interactions have been moved to a separate class. (#7)"
This reverts commit 484131675432f9e9951362f315b09aa1534ddc25.
