+++ HumanResourceMachine/Controllers/HumanController.cs
@@ -1,6 +1,5 @@
 ﻿using HumanResourceMachine.Context;
 using HumanResourceMachine.Entities;
-using HumanResourceMachine.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -11,41 +10,56 @@ namespace HumanResourceMachine.Controllers
     [ApiController]
     public class HumanController : ControllerBase
     {
-        private readonly IBusinessLogicService _service;
+        private readonly HRMContext _context;
 
-        public HumanController(IBusinessLogicService service)
+        public HumanController(HRMContext context)
         {
-            _service = service;
+            _context = context;
         }
 
         [HttpGet]
         public IEnumerable<Human> GetAllHumans()
         {
-            return _service.GetAllHumans();
+            return _context.People.ToList();
         }
 
         [HttpGet("{id}")]
         public Human GetHumanById([FromRoute] int id)
         {
-            return _service.GetHumanById(id);
+            var result = _context.People.Find(id);
+            if (result is null)
+            {
+                throw new ArgumentNullException("Object doesn't exist.", nameof(result));
+            }
+
+            return result;
         }
 
         [HttpPost]
         public void AddHuman(Human human)
         {
-            _service.AddHuman(human);
+            _context.People.Add(human);
+            _context.SaveChanges();
         }
 
         [HttpPut]
         public void UpdateHuman(Human human)    
         {
-            _service.UpdateHuman(human);
+            _context.Update(human); 
+            _context.SaveChanges();
         }
 
         [HttpDelete("{id}")]
         public void DeleteHumanById([FromRoute] int id)
         {
-            _service.DeleteHumanById(id);
+            var target = _context.People.Find(id);
+            if (target is null)
+            {
+                throw new ArgumentNullException("Object doesn't exist.", nameof(target));
+            }
+
+            _context.People.Remove(target); 
+            _context.SaveChanges();
         }
     }
 }
