+++ HumanResourceMachine/Program.cs
@@ -1,7 +1,5 @@
 using HumanResourceMachine.Context;
-using HumanResourceMachine.Interfaces;
 using HumanResourceMachine.Middleware;
-using HumanResourceMachine.Services;
 using Microsoft.EntityFrameworkCore;
 
 namespace HumanResourceMachine
@@ -15,7 +13,6 @@ namespace HumanResourceMachine
             // Add services to the container.
             string connection = builder.Configuration.GetConnectionString("DefaultConnection");
             builder.Services.AddDbContext<HRMContext>(options => options.UseSqlServer(connection));
-            builder.Services.AddTransient<IBusinessLogicService, BusinessLogicService>();
             builder.Services.AddControllers();
             // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
             builder.Services.AddEndpointsApiExplorer();
